split analysis into submodule
